requirements.txt
streamlit==1.3.1
ta==0.7.0
python-binance==1.0.15
matplotlib==3.5.0
ccxt==1.59.95
scikit-learn==1.0
cryptography==3.1.1
yfinance==0.1.64